Refactor: packages optimization(dto, service)
