+++ src/main/java/mate/academy/bookstore/controller/AuthenticationController.java
@@ -4,10 +4,10 @@ import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
-import mate.academy.bookstore.dto.UserLoginRequestDto;
-import mate.academy.bookstore.dto.UserLoginResponseDto;
-import mate.academy.bookstore.dto.UserRegistrationRequestDto;
-import mate.academy.bookstore.dto.UserResponseDto;
+import mate.academy.bookstore.dto.user.UserLoginRequestDto;
+import mate.academy.bookstore.dto.user.UserLoginResponseDto;
+import mate.academy.bookstore.dto.user.UserRegistrationRequestDto;
+import mate.academy.bookstore.dto.user.UserResponseDto;
 import mate.academy.bookstore.exception.RegistrationException;
 import mate.academy.bookstore.security.AuthenticationService;
 import mate.academy.bookstore.service.UserService;
