+++ src/main/java/mate/academy/bookstore/dto/book/BookSearchParametersDto.java
@@ -1,4 +1,4 @@
-package mate.academy.bookstore.dto;
+package mate.academy.bookstore.dto.book;
 
 public record BookSearchParametersDto(
         String title,
