+++ src/main/java/mate/academy/bookstore/dto/user/UserLoginResponseDto.java
@@ -1,4 +1,4 @@
-package mate.academy.bookstore.dto;
+package mate.academy.bookstore.dto.user;
 
 public record UserLoginResponseDto(String token) {
 }
