+++ src/main/java/mate/academy/bookstore/repository/SpecificationBuilder.java
@@ -1,6 +1,6 @@
 package mate.academy.bookstore.repository;
 
-import mate.academy.bookstore.dto.BookSearchParametersDto;
+import mate.academy.bookstore.dto.book.BookSearchParametersDto;
 import org.springframework.data.jpa.domain.Specification;
 
 public interface SpecificationBuilder<T> {
