+++ src/main/java/mate/academy/bookstore/security/AuthenticationService.java
@@ -1,8 +1,8 @@
 package mate.academy.bookstore.security;
 
 import lombok.RequiredArgsConstructor;
-import mate.academy.bookstore.dto.UserLoginRequestDto;
-import mate.academy.bookstore.dto.UserLoginResponseDto;
+import mate.academy.bookstore.dto.user.UserLoginRequestDto;
+import mate.academy.bookstore.dto.user.UserLoginResponseDto;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
