+++ src/main/java/mate/academy/bookstore/service/impl/CategoryServiceImpl.java
@@ -1,4 +1,4 @@
-package mate.academy.bookstore.service.category;
+package mate.academy.bookstore.service.impl;
 
 import lombok.RequiredArgsConstructor;
 import mate.academy.bookstore.dto.category.CategoryDto;
@@ -7,6 +7,7 @@ import mate.academy.bookstore.exception.EntityNotFoundException;
 import mate.academy.bookstore.mapper.CategoryMapper;
 import mate.academy.bookstore.model.Category;
 import mate.academy.bookstore.repository.category.CategoryRepository;
+import mate.academy.bookstore.service.CategoryService;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
